requirements.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
certifi==2017.7.27.1
chardet==3.0.4
django==1.11.3
django-enumfields==0.9.0
djangorestframework==3.6.3
idna==2.5
psycopg2==2.7.3
pytz==2017.2
requests==2.18.2
urllib3==1.22
